fix(mobile): address PR feedback for Checkbox component bb-174
