+++ apps/mobile/src/libs/enums/app/app.ts
@@ -2,6 +2,5 @@ export { AngleGradient } from "./angle-gradient.enum";
 export { BaseColor } from "./base-color.enum";
 export { DataStatus } from "./data-status.enum";
 export { GradientColor } from "./gradient-color.enum";
-export { GradientPoint } from "./gradient-point.enum";
 export { ToastMessageTitle } from "./toast-message-title.enum";
 export { ToastMessageType } from "./toast-message-type.enum";
