--- apps/mobile/src/libs/enums/app/gradient-point.enum.ts
@@ -1,8 +1,0 @@
-const GradientPoint = {
-	ANGLE_CENTER_POINT: 0.5,
-	FIRST_COLOR_STOP: -0.4,
-	GRADIENT_ANGLE: 305,
-	SECOND_COLOR_STOP: 0.9,
-} as const;
-
-export { GradientPoint };
