+++ apps/mobile/src/libs/enums/navigation/root-screen-name.enum.ts
@@ -1,6 +1,6 @@
 const RootScreenName = {
-  CHAT: "Chat",
 	BOTTOM_TABS_NAVIGATOR: "BottomTabsNavigator",
+	CHAT: "Chat",
 	QUIZ_ENTRY: "Quiz Entry",
 	SIGN_IN: "Sign In",
 	SIGN_UP: "Sign Up",
