+++ apps/mobile/src/libs/types/navigation/root-navigation-parameter-list.type.ts
@@ -1,8 +1,8 @@
 import { type RootScreenName } from "~/libs/enums/enums";
 
 type RootNavigationParameterList = {
-	[RootScreenName.CHAT]: undefined;
 	[RootScreenName.BOTTOM_TABS_NAVIGATOR]: undefined;
+	[RootScreenName.CHAT]: undefined;
 	[RootScreenName.QUIZ_ENTRY]: undefined;
 	[RootScreenName.SIGN_IN]: undefined;
 	[RootScreenName.SIGN_UP]: undefined;
